fix dropdown in mobile view
